Solving warnings
