+++ WolmoCoreDemo/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -103,8 +103,9 @@
       "scale" : "2x"
     },
     {
-      "idiom" : "ios-marketing",
       "size" : "1024x1024",
+      "idiom" : "ios-marketing",
+      "filename" : "Icon-1024.png",
       "scale" : "1x"
     }
   ],
